apps/monitor-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
watchdog==5.0.3
paramiko==4.0.0
python-dotenv==1.0.1
pydantic==2.9.2
asyncio-mqtt==0.16.2
